일기 작성 페이지 Container, Component 분리
