+++ schedule.md
@@ -36,6 +36,7 @@
 -   헤더 구현 & 메인 페이지 디자인 입히기
 
 [2021.10.13]
+3h
 
 ## 계획
 
@@ -57,3 +58,30 @@
 -   firebase api 가져오는 부분 정리하기
     컬렉션 목록 가져올때 자동생성되는 id를 도큐먼트리스트안에 넣어서 가져오는 방법. (타입스크립트)
 -   날짜, 입력일자로 정렬해야함. 입력일자를 insert 하도록
+
+[2021.10.15]
+2h
+
+## 실제한일
+
+-   도메인 구입 후 연결(http://jimin.xyz:50091/)
+
+## 할일
+
+-   도메인 포트번호 없애는 방법 알아보기.
+
+[2021.10.16]
+
+-   작업 없음
+
+[2021.10.17]
+
+## 예정
+
+-   nginx 서버
+-   리펙토링
+-   날짜, 입력일자 받기
+-   텍스트 에디터 추가하기
+-   반응형 적용하기
+-   1차 릴리즈 태그 추가하기
+-   github action 이용하여 자동 배포 설정하기
