+++ src/containers/diary/DiaryDetailLoader.tsx
@@ -4,7 +4,7 @@ import { RootState } from '../../modules';
 import { getDiaryDetailAsync } from '../../modules/diary';
 import DiaryViewer from '../../components/diary/DiaryViewer';
 import styled from 'styled-components';
-import Loading from '../../components/loading/Loadingg';
+import Loading from '../../components/loading/Loading';
 
 type DiaryDetailLoaderProps = {
     id: string;
