+++ src/lib/utils.ts
@@ -7,21 +7,21 @@ export const formatDate = (date: string): string => {
     const now = Date.now();
     const diff = now - new Date(date).getTime();
 
-    if (diff < 1000 * 60 * 5) {
-        return '방금 전';
-    }
+    // if (diff < 1000 * 60 * 5) {
+    //     return '방금 전';
+    // }
 
-    if (diff < 1000 * 60 * 60 * 24) {
-        return distanceInWordsToNow(d, { addSuffix: true, locale: koLocale });
-    }
+    // if (diff < 1000 * 60 * 60 * 24) {
+    //     return distanceInWordsToNow(d, { addSuffix: true, locale: koLocale });
+    // }
 
-    if (diff < 1000 * 60 * 60 * 36) {
-        return '어제';
-    }
+    // if (diff < 1000 * 60 * 60 * 36) {
+    //     return '어제';
+    // }
 
-    if (diff < 1000 * 60 * 60 * 24 * 7) {
-        return distanceInWordsToNow(d, { addSuffix: true, locale: koLocale });
-    }
+    // if (diff < 1000 * 60 * 60 * 24 * 7) {
+    //     return distanceInWordsToNow(d, { addSuffix: true, locale: koLocale });
+    // }
     return format(d, 'yyyy년 M월 d일');
 };
